+++ src/datalab_jupyter/server/__init__.py
@@ -1,1 +1,1 @@
-"""Jupyter Server integration for Datalab user environments."""
+"""Jupyter Server integration for datalab user environments."""